+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Add child logger with instance named tags support
+
 ## [4.17.0]
 
 - Correct `source_code_uri` URL